upgrade miniz and use inline
