+++ contrib/miniz/Makefile.am
@@ -1,6 +1,2 @@
 #  Process this file with automake to produce Makefile.in
 include $(top_srcdir)/Makefile.common
-
-noinst_LIBRARIES = libminiz.a
-libminiz_a_SOURCES = miniz.c
-noinst_HEADERS = miniz.h
